+++ lib/src/builder/reflection_writer.dart
@@ -1,6 +1,5 @@
-import 'package:analyzer/dart/ast/ast.dart';
-import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/nullability_suffix.dart';
 
 import 'annotations.dart';
 import 'class_finders.dart';
@@ -43,17 +42,22 @@ class ReflectionWriter {
 
     sb.write("    "); //lvl 2 indent
     if (p.isPositional) {
-      _writeWrapped(convFun, () => attribWriter(p));
+      _writeWrapped(p, convFun, () => attribWriter(p));
       sb.writeln(",");
     } else {
       sb.write(p.name);
       sb.write(": ");
-      _writeWrapped(convFun, () => attribWriter(p));
+      _writeWrapped(p, convFun, () => attribWriter(p));
       sb.writeln(",");
     }
   }
 
-  void _writeWrapped(ConvertFunction? convFun, VoidFunction callback) {
+  void _writeWrapped(
+      ParameterElement p, ConvertFunction? convFun, VoidFunction callback) {
+    if (p.name == "child" || p.name == "children") {
+      callback();
+      return;
+    }
     if (convFun != null) {
       sb.write(convFun.functionName);
       sb.write("(");
@@ -61,9 +65,26 @@ class ReflectionWriter {
 
     callback();
 
+    bool canBeNull = p.type.nullabilitySuffix == NullabilitySuffix.question;
     if (convFun != null) {
+      if (convFun.nullableResult) {
+        sb.write(" ?? ''");
+      }
       sb.write(")");
       if (convFun.nullableResult) {
+        if (p.hasDefaultValue) {
+          sb.write(" ?? ");
+          sb.write(p.defaultValueCode);
+        } else if (!canBeNull) {
+          sb.write("!");
+        }
+      }
+    } else {
+      //No convert function, but still need to check
+      if (p.hasDefaultValue) {
+        sb.write(" ?? ");
+        sb.write(p.defaultValueCode);
+      } else if (!canBeNull) {
         sb.write("!");
       }
     }
@@ -78,11 +99,15 @@ class ReflectionWriter {
     //Check if expected param type is different then String, if yes perform
     //conversion
     if (p.type.isDartCoreInt) {
-      return ConvertFunction.withFunc("int.parse", false, []);
+      return p.type.nullabilitySuffix == NullabilitySuffix.question
+          ? ConvertFunction.withFunc("int.tryParse", true, [])
+          : ConvertFunction.withFunc("int.parse", false, []);
     } else if (p.type.isDartCoreDouble) {
-      return ConvertFunction.withFunc("double.parse", false, []);
+      return p.type.nullabilitySuffix == NullabilitySuffix.question
+          ? ConvertFunction.withFunc("double.tryParse", true, [])
+          : ConvertFunction.withFunc("double.parse", false, []);
     } else if (p.type.isDartCoreBool) {
-      return ConvertFunction.withFunc("bool.parse", false, []);
+      return ConvertFunction.withFunc("'true' == ", false, []);
     } else if (!p.type.isDartCoreString) {
       print("Probably this will cause problems, $p");
     }